+++ pkg/providers/appsec/create_appsec_test.go
@@ -119,7 +119,9 @@ func TestExportRatePolicy(t *testing.T) {
     "additionalMatchOptions": null,
     "averageThreshold": 100,
     "burstThreshold": 500,
-    "clientIdentifier": "ip",
+    "clientIdentifiers": [
+        "ip"
+    ],
     "matchType": "path",
     "name": "High Rate",
     "pathMatchType": "Custom",
@@ -134,7 +136,9 @@ func TestExportRatePolicy(t *testing.T) {
     "additionalMatchOptions": null,
     "averageThreshold": 100,
     "burstThreshold": 500,
-    "clientIdentifier": "ip",
+    "clientIdentifiers": [
+        "ip"
+    ],
     "matchType": "path",
     "name": "High Rate",
     "pathMatchType": "Custom",