+++ pkg/providers/appsec/testdata/ase/modules/security/rate-policies.tf
@@ -14,7 +14,9 @@ resource "akamai_appsec_rate_policy" "page_view_requests" {
       ],
       "averageThreshold" : 12,
       "burstThreshold" : 18,
-      "clientIdentifier" : "ip",
+      "clientIdentifiers" : [
+        "ip"
+      ],
       "description" : "A popular brute force attack that consists of sending a large number of requests for base page, HTML page or XHR requests (usually non-cacheable). This could destabilize the origin.",
       "fileExtensions" : {
         "positiveMatch" : false,
@@ -124,7 +126,9 @@ resource "akamai_appsec_rate_policy" "origin_error" {
       ],
       "averageThreshold" : 5,
       "burstThreshold" : 8,
-      "clientIdentifier" : "ip",
+      "clientIdentifiers" : [
+        "ip"
+      ],
       "description" : "An excessive error rate from the origin could indicate malicious activity by a bot scanning the site or a publishing error. In both cases, this would increase the origin traffic and could potentially destabilize it.",
       "matchType" : "path",
       "name" : "Origin Error",
@@ -153,7 +157,9 @@ resource "akamai_appsec_rate_policy" "post_page_requests" {
       ],
       "averageThreshold" : 3,
       "burstThreshold" : 5,
-      "clientIdentifier" : "ip",
+      "clientIdentifiers" : [
+        "ip"
+      ],
       "description" : "Mitigating HTTP flood attacks using POST requests",
       "matchType" : "path",
       "name" : "POST Page Requests",